pyproject.toml
[tool.poetry]
name = "user-registration"
version = "0.1.0"
description = ""
authors = ["SANCHAY"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
pydantic = "^2.10.4"
uvicorn = "^0.34.0"
asyncpg = "^0.30.0"
bcrypt = "^4.2.1"
jwt = "^1.3.1"
pyjwt = "^2.10.1"
python-jose = "^3.3.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"  #


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
